Moved object initializations within constructor and created a separate package for the Course class
